feat(users): separate update command
